% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.numSummary.R
\name{plot.numSummary}
\alias{plot.numSummary}
\title{Plot numeric summary with ggplot2}
\usage{
\method{plot}{numSummary}(obj)
}
\arguments{
\item{obj}{A numSummary object.}
}
\value{
A ggplot2 bar chart of summary statistics.
}
\description{
Creates a ggplot2 bar chart of mean, min, max, and median
from a numSummary object.
}
\examples{
x <- c(1, 5, 3, NA, 9, 2)
s <- numSummary(x)
plot(s)
}
