% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.numSummary.R
\name{print.numSummary}
\alias{print.numSummary}
\title{Print method for numSummary objects}
\usage{
\method{print}{numSummary}(obj)
}
\arguments{
\item{obj}{A numSummary object}
}
\description{
Print method for numSummary objects
}
